requirements.txt
Django
pyrebase4
django-cors-headers
djangorestframework